+++ unstructured/partition/common.py
@@ -1,5 +1,7 @@
 import subprocess
-from typing import List, Optional, Tuple, Union
+from io import BytesIO
+from tempfile import SpooledTemporaryFile
+from typing import BinaryIO, List, Optional, Tuple, Union
 
 from unstructured.documents.elements import (
     TYPE_TO_TEXT_ELEMENT_MAP,
@@ -16,6 +18,9 @@ from unstructured.nlp.patterns import ENUMERATED_BULLETS_RE, UNICODE_BULLETS_RE
 def normalize_layout_element(layout_element) -> Union[Element, List[Element]]:
     """Converts a list of unstructured_inference DocumentLayout objects to a list of
     unstructured Elements."""
+
+    if isinstance(layout_element, Element):
+        return layout_element
 
     if isinstance(layout_element, PageBreak):
         return PageBreak()
@@ -62,26 +67,11 @@ def layout_list_to_list_items(
     return list_items
 
 
-def document_to_element_list(
-    document,
-    include_page_breaks: bool = False,
-) -> List[Element]:
-    """Converts a DocumentLayout object to a list of unstructured elements."""
-    elements: List[Element] = []
-    num_pages = len(document.pages)
-    for i, page in enumerate(document.pages):
-        for element in page.elements:
-            elements.append(element)
-        if include_page_breaks and i < num_pages - 1:
-            elements.append(PageBreak())
-
-    return elements
-
-
-def add_element_metadata(
+def _add_element_metadata(
     layout_elements,
     include_page_breaks: bool = False,
     filename: Optional[str] = None,
+    filetype: Optional[str] = None,
     url: Optional[str] = None,
 ) -> List[Element]:
     """Adds document metadata to the document element. Document metadata includes information
@@ -96,20 +86,21 @@ def add_element_metadata(
             text_as_html = None
         metadata = ElementMetadata(
             filename=filename,
+            filetype=filetype,
             url=url,
             page_number=page_number,
             text_as_html=text_as_html,
         )
         if isinstance(element, list):
             for _element in element:
-                _element.metadata = metadata
+                _element.metadata = metadata.merge(_element.metadata)
             elements.extend(element)
         elif isinstance(element, PageBreak):
             page_number += 1
-            if include_page_breaks is True:
+            if include_page_breaks:
                 elements.append(element)
         else:
-            element.metadata = metadata
+            element.metadata = metadata.merge(element.metadata)
             elements.append(element)
     return elements
 
@@ -157,3 +148,15 @@ def exactly_one(**kwargs) -> None:
         else:
             message = f"{names[0]} must be specified."
         raise ValueError(message)
+
+
+def spooled_to_bytes_io_if_needed(
+    file_obj: Optional[Union[bytes, BinaryIO, SpooledTemporaryFile]],
+) -> Optional[Union[bytes, BinaryIO]]:
+    if isinstance(file_obj, SpooledTemporaryFile):
+        file_obj.seek(0)
+        contents = file_obj.read()
+        return BytesIO(contents)
+    else:
+        # Return the original file object if it's not a SpooledTemporaryFile
+        return file_obj