+++ unstructured/partition/pdf.py
@@ -1,33 +1,40 @@
 import re
 import warnings
-from typing import BinaryIO, List, Optional, cast
+from tempfile import SpooledTemporaryFile
+from typing import BinaryIO, List, Optional, Union, cast
 
+import pdf2image
 from pdfminer.high_level import extract_pages
-from pdfminer.pdfpage import PDFPage, PDFTextExtractionNotAllowed
 from pdfminer.utils import open_filename
+from PIL import Image
 
 from unstructured.cleaners.core import clean_extra_whitespace
 from unstructured.documents.elements import Element, ElementMetadata, PageBreak
-from unstructured.logger import logger
+from unstructured.file_utils.filetype import (
+    FileType,
+    add_metadata_with_filetype,
+    document_to_element_list,
+)
 from unstructured.nlp.patterns import PARAGRAPH_PATTERN
 from unstructured.partition import _partition_via_api
 from unstructured.partition.common import (
-    add_element_metadata,
-    document_to_element_list,
     exactly_one,
+    spooled_to_bytes_io_if_needed,
 )
+from unstructured.partition.strategies import determine_pdf_or_image_strategy
 from unstructured.partition.text import partition_text
-from unstructured.utils import dependency_exists, requires_dependencies
+from unstructured.utils import requires_dependencies
 
 
+@add_metadata_with_filetype(FileType.PDF)
 def partition_pdf(
     filename: str = "",
-    file: Optional[bytes] = None,
+    file: Optional[Union[BinaryIO, SpooledTemporaryFile]] = None,
     url: Optional[str] = None,
     template: str = "layout/pdf",
     token: Optional[str] = None,
     include_page_breaks: bool = False,
-    strategy: str = "hi_res",
+    strategy: str = "auto",
     infer_table_structure: bool = False,
     encoding: str = "utf-8",
     ocr_languages: str = "eng",
@@ -48,9 +55,12 @@ def partition_pdf(
     token
         A string defining the authentication token for a self-host url, if applicable.
     strategy
-        The strategy to use for partitioning the PDF. Uses a layout detection model if set
-        to 'hi_res', otherwise partition_pdf simply extracts the text from the document
-        and processes it.
+        The strategy to use for partitioning the PDF. Valid strategies are "hi_res",
+        "ocr_only", and "fast". When using the "hi_res" strategy, the function uses
+        a layout detection model to identify document elements. When using the
+        "ocr_only" strategy, partition_image simply extracts the text from the
+        document using OCR and processes it. If the "fast" strategy is used, the text
+        is extracted directly from the PDF.
     infer_table_structure
         Only applicable if `strategy=hi_res`.
         If True, any Table elements that are extracted will also have a metadata field
@@ -81,13 +91,13 @@ def partition_pdf(
 
 def partition_pdf_or_image(
     filename: str = "",
-    file: Optional[bytes] = None,
+    file: Optional[Union[bytes, BinaryIO, SpooledTemporaryFile]] = None,
     url: Optional[str] = "https://ml.unstructured.io/",
     template: str = "layout/pdf",
     token: Optional[str] = None,
     is_image: bool = False,
     include_page_breaks: bool = False,
-    strategy: str = "hi_res",
+    strategy: str = "auto",
     infer_table_structure: bool = False,
     encoding: str = "utf-8",
     ocr_languages: str = "eng",
@@ -104,42 +114,21 @@ def partition_pdf_or_image(
         if route_args[0] == "layout":
             out_template = None
 
-        fallback_to_fast = False
-        fallback_to_hi_res = False
+        strategy = determine_pdf_or_image_strategy(
+            strategy,
+            filename=filename,
+            file=file,
+            is_image=is_image,
+            infer_table_structure=infer_table_structure,
+        )
 
-        detectron2_installed = dependency_exists("detectron2")
-        if is_image:
-            pdf_text_extractable = False
-        else:
-            pdf_text_extractable = is_pdf_text_extractable(filename=filename, file=file)
-            if file is not None:
-                file.seek(0)  # type: ignore
-
-        if not detectron2_installed and not pdf_text_extractable:
-            raise ValueError(
-                "detectron2 is not installed and the text of the PDF is not extractable. "
-                "To process this file, install detectron2 or remove copy protection from the PDF.",
-            )
-
-        if not pdf_text_extractable:
-            fallback_to_hi_res = strategy == "fast"
-
-        if not detectron2_installed:
-            fallback_to_fast = strategy == "hi_res"
-
-        if (strategy == "hi_res" or fallback_to_hi_res) and not fallback_to_fast:
-            if strategy == "fast":
-                logger.warning(
-                    "PDF text is not extractable. Cannot use the fast partitioning "
-                    "strategy. Falling back to partitioning with the hi_res strategy.",
-                )
-
+        if strategy == "hi_res":
             # NOTE(robinson): Catches a UserWarning that occurs when detectron is called
             with warnings.catch_warnings():
                 warnings.simplefilter("ignore")
                 layout_elements = _partition_pdf_or_image_local(
                     filename=filename,
-                    file=file,
+                    file=spooled_to_bytes_io_if_needed(file),
                     template=out_template,
                     is_image=is_image,
                     infer_table_structure=infer_table_structure,
@@ -147,27 +136,24 @@ def partition_pdf_or_image(
                     ocr_languages=ocr_languages,
                 )
 
-        elif (strategy == "fast" or fallback_to_fast) and not fallback_to_hi_res:
-            if strategy == "hi_res":
-                logger.warning(
-                    "detectron2 is not installed. Cannot use the hi_res partitioning "
-                    "strategy. Falling back to partitioning with the fast strategy.",
-                )
-            if infer_table_structure:
-                logger.warning(
-                    "Table extraction was selected, but is being ignored while using the fast "
-                    "strategy.",
-                )
-
+        elif strategy == "fast":
             return _partition_pdf_with_pdfminer(
                 filename=filename,
-                file=file,
+                file=spooled_to_bytes_io_if_needed(file),
                 include_page_breaks=include_page_breaks,
                 encoding=encoding,
             )
 
-        else:
-            raise ValueError(f"{strategy} is an invalid parsing strategy for PDFs")
+        elif strategy == "ocr_only":
+            # NOTE(robinson): Catches file conversion warnings when running with PDFs
+            with warnings.catch_warnings():
+                return _partition_pdf_or_image_with_ocr(
+                    filename=filename,
+                    file=file,
+                    include_page_breaks=include_page_breaks,
+                    ocr_languages=ocr_languages,
+                    is_image=is_image,
+                )
 
     else:
         # NOTE(alan): Remove these lines after different models are handled by routing
@@ -179,23 +165,20 @@ def partition_pdf_or_image(
         # NOTE(alan): Remove "data=data" after different models are handled by routing
         layout_elements = _partition_via_api(
             filename=filename,
-            file=file,
+            file=cast(BinaryIO, file),
             url=url,
             token=token,
             data=data,
             include_page_breaks=True,
         )
 
-    return add_element_metadata(
-        layout_elements,
-        include_page_breaks=include_page_breaks,
-        filename=filename,
-    )
+    return layout_elements
 
 
+@requires_dependencies("unstructured_inference")
 def _partition_pdf_or_image_local(
     filename: str = "",
-    file: Optional[bytes] = None,
+    file: Optional[Union[bytes, BinaryIO]] = None,
     template: Optional[str] = None,
     is_image: bool = False,
     infer_table_structure: bool = False,
@@ -246,7 +229,7 @@ def _partition_pdf_or_image_local(
 @requires_dependencies("pdfminer", "local-inference")
 def _partition_pdf_with_pdfminer(
     filename: str = "",
-    file: Optional[bytes] = None,
+    file: Optional[BinaryIO] = None,
     include_page_breaks: bool = False,
     encoding: str = "utf-8",
 ) -> List[Element]:
@@ -318,23 +301,42 @@ def _process_pdfminer_pages(
     return elements
 
 
-def is_pdf_text_extractable(filename: str = "", file: Optional[bytes] = None):
-    """Checks to see if the text from a PDF document is extractable. Sometimes the
-    text is not extractable due to PDF security settings."""
-    exactly_one(filename=filename, file=file)
+@requires_dependencies("pytesseract")
+def _partition_pdf_or_image_with_ocr(
+    filename: str = "",
+    file: Optional[Union[bytes, BinaryIO, SpooledTemporaryFile]] = None,
+    include_page_breaks: bool = False,
+    ocr_languages: str = "eng",
+    is_image: bool = False,
+):
+    """Partitions and image or PDF using Tesseract OCR. For PDFs, each page is converted
+    to an image prior to processing."""
+    import pytesseract
 
-    def _fp_is_extractable(fp):
-        try:
-            next(PDFPage.get_pages(fp, check_extractable=True))
-            extractable = True
-        except PDFTextExtractionNotAllowed:
-            extractable = False
-        return extractable
+    if is_image:
+        if file is not None:
+            image = Image.open(file)
+            text = pytesseract.image_to_string(image, config=f"-l '{ocr_languages}'")
+        else:
+            text = pytesseract.image_to_string(filename, config=f"-l '{ocr_languages}'")
+        elements = partition_text(text=text)
+    else:
+        elements = []
+        if file is not None:
+            document = pdf2image.convert_from_bytes(file.read())  # type: ignore
+            file.seek(0)  # type: ignore
+        else:
+            document = pdf2image.convert_from_path(filename)
 
-    if filename:
-        with open_filename(filename, "rb") as fp:
-            fp = cast(BinaryIO, fp)
-            return _fp_is_extractable(fp)
-    elif file:
-        fp = cast(BinaryIO, file)
-        return _fp_is_extractable(fp)
+        for i, image in enumerate(document):
+            metadata = ElementMetadata(filename=filename, page_number=i + 1)
+            text = pytesseract.image_to_string(image, config=f"-l '{ocr_languages}'")
+
+            _elements = partition_text(text=text)
+            for element in _elements:
+                element.metadata = metadata
+                elements.append(element)
+
+            if include_page_breaks:
+                elements.append(PageBreak())
+    return elements